Create Bookshop management system
It's a try to build a book shop management system to keep a record of books in a shop. This has been approached by connecting the cpp code to database. via postgresql. The code includes the following functions,  connecting to the db, creating new record of the book, search for the book, update the details of the book, deleting the record and showing the details of the records.